# resources/swagger.yaml
openapi: 3.0.0
info:
  title: Modelo Movelaria API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /orcamentos:
    post:
      summary: Criar orçamento
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrcamentoCreate'
      responses:
        '201':
          description: Orçamento criado
    get:
      summary: Listar orçamentos
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Lista de orçamentos
  /orcamentos/{id}:
    get:
      summary: Obter orçamento por ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Orçamento
    put:
      summary: Atualizar orçamento (campos camelCase; ambientes substitui a lista; status Fechado gera o cliente)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrcamentoUpdate'
      responses:
        '204':
          description: Atualizado
...
  /ambientes/orcamento/{id}:
    get:
      summary: Listar ambientes por orçamento
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de ambientes
...
  /arquivos/orcamento/{id}:
    get:
      summary: Listar arquivos de um orçamento
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de arquivos
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrcamentoCreate:
      type: object
      properties:
        IDContato:
          type: integer
        idProspeccao:
          type: integer
          description: Prospecção que originou o orçamento. Ao criar, ela passa a Convertido e fica vinculada (409 se já gerou orçamento).
        IDStatusOrcamento:
          type: integer
        IDFuncionario:
          type: integer
        NumeroOrcamento:
          type: string
        DataSolicitacao:
          type: string
          format: date
        NecessitaProjeto:
          type: boolean
        NecessitaVisita:
          type: boolean
        Observacao:
          type: string
        Valor:
          type: number
        DataEntregaOrcamento:
          type: string
          format: date
        IDEmpresa:
          type: integer
    OrcamentoUpdate:
      type: object
      properties:
        Observacao:
          type: string
        Valor:
          type: number
        DataFechamento:
          type: string
          format: date
...